+++ src/layers/CreatorLayer.cpp
@@ -9,9 +9,11 @@ class $modify(RouletteButton, CreatorLayer)
 {
 	void onRouletteButton(CCObject*)
 	{
-		g_rouletteManager.rouletteLayer = RLRouletteLayer::create();
-		if (g_rouletteManager.rouletteLayer)
+		if (g_rouletteManager.rouletteLayer = RLRouletteLayer::create(); g_rouletteManager.rouletteLayer)
 			g_rouletteManager.rouletteLayer->show();
+
+		if (auto exMark = this->getChildByID("center-left-menu"_spr)->getChildByID("roulette-button"_spr)->getChildByID("exclamation-mark"))
+			exMark->removeFromParent();
 	}
 
 	bool init()
@@ -20,34 +22,28 @@ class $modify(RouletteButton, CreatorLayer)
 
 		auto winSize = CCDirector::sharedDirector()->getWinSize();
 
-		if(Mod::get()->getSettingValue<bool>("use-big-button") && Loader::get()->isModLoaded("alphalaneous.pages_api")) {
-			if(auto menu = getChildByID("creator-buttons-menu")) {
-				auto spr = CCSprite::create("RL_btn_big.png"_spr);
-				spr->setScale(0.842);
-				auto rouletteButton = CCMenuItemSpriteExtra::create(
-					spr,
-					this,
-					menu_selector(RouletteButton::onRouletteButton)
-				);
-				rouletteButton->setID("roulette-button"_spr);
-				menu->addChild(rouletteButton);
-			}
-		}
-		else {
-			if(auto menu = getChildByID("exit-menu")) {
-				auto spr = CircleButtonSprite::createWithSprite("RL_btn_001.png"_spr, 1.0f, CircleBaseColor::Green, CircleBaseSize::Small);
+		auto centerLeftMenu = CCMenu::create();
+		centerLeftMenu->setPosition({
+			this->getChildByID("exit-menu")->getPositionX(),
+			CCDirector::sharedDirector()->getWinSize().height / 2.f
+		});
+		centerLeftMenu->setID("center-left-menu"_spr);
+		this->addChild(centerLeftMenu);
 
-				auto rouletteButton = CCMenuItemSpriteExtra::create(
-					spr,
-					this,
-					menu_selector(RouletteButton::onRouletteButton)
-				);
-				rouletteButton->setID("roulette-button"_spr);
-				menu->addChild(rouletteButton);
-				typeinfo_cast<AxisLayout*>(menu->getLayout())->setAxisReverse(true);
-				menu->updateLayout();
-			}
-		}
+		auto spr = CircleButtonSprite::createWithSprite("RL_btn_001.png"_spr, 1.0f, CircleBaseColor::Green, CircleBaseSize::Small);
+		spr->setScale(1.1f);
+
+		auto rouletteButton = CCMenuItemSpriteExtra::create(
+			spr,
+			this,
+			menu_selector(RouletteButton::onRouletteButton)
+		);
+		rouletteButton->setID("roulette-button"_spr);
+		rouletteButton->setLayout(AnchorLayout::create());
+		centerLeftMenu->addChild(rouletteButton);
+
+		if (g_rouletteManager.gameState.levelID != 0)
+			g_rouletteManager.addExclamationMark(rouletteButton);
 
 		return true;
 	}