+++ src/roulette/manager/RouletteManager.hpp
@@ -5,6 +5,8 @@
 #include <Geode/loader/Mod.hpp>
 #include <Geode/Enums.hpp>
 
+#include "DataManager.hpp"
+#include "GameState.hpp"
 #include "../layers/RLRouletteLayer.hpp"
 #include "../../utils.hpp"
 
@@ -16,59 +18,15 @@ struct RouletteManager
 private:
 	inline static bool m_hasInitManager = false;
 
-	struct SavedArrayInfo
-	{
-		const std::size_t size;
-		const std::vector<bool> default_value;
-
-		SavedArrayInfo(std::size_t size, const std::vector<bool>& default_value)
-			: size(size), default_value(default_value)
-		{}
-	};
-
-	const std::unordered_map<std::string, SavedArrayInfo> ARRAY_TO_SAI{
-		{ "difficulty-array", { 6, { true, false, false, false, false, false } } },
-		{ "demon-difficulty-array", { 5, { true, false, false, false, false, false }} },
-		{ "selected-list-array", { 4, { true, false, false, false }} }
-	};
-
-	void verifyArray(const std::string& name, matjson::Array& arr)
-	{
-		if (rl::utils::getCountOf(arr, true) != 1)
-		{
-			for (auto& elem : arr)
-				elem = false;
-			arr.at(0) = true;
-		}
-
-		if (
-			auto arraySaveDataInfo = ARRAY_TO_SAI.at(name);
-			arraySaveDataInfo.size != arr.size()
-		) {
-			auto previousIndex = rl::utils::getIndexOf(arr, true);
-
-			Mod::get()->setSavedValue<std::vector<bool>>(name, arraySaveDataInfo.default_value);
-
-			if (previousIndex != -1)
-				getFromSaveContainer(name).as_array().at(previousIndex) = true;
-		}
-	}
-
 public:
 	inline static RLRouletteLayer* rouletteLayer;
 
 	inline static std::atomic_bool isPlaying = false;
 	inline static std::atomic_bool isPaused = false;
-	inline static bool hasFinishedPreviousLevel = false;
 
-	inline static int currentLevelID = 0;
-	inline static int currentLevelPercentage = 0;
-	inline static int levelPercentageGoal = 1;
-
-	inline static int skipsUsed = 0;
-
-	inline static int numLevels = 0;
-
+	inline static GameState gameState{};
+	// used visually, not saved.
+	inline static int currentPercentageGoal = 1;
 	inline static int gdListID = 0;
 
 	inline static GJDifficulty previousDifficulty = GJDifficulty::Easy;
@@ -79,55 +37,54 @@ public:
 	{
 		if (!m_hasInitManager)
 		{
-			auto& saveContainer = Mod::get()->getSaveContainer().as_object();
+			auto& saveContainer = Mod::get()->getSaveContainer();
 
-			if (!Mod::get()->hasSavedValue("difficulty-array"))
-				Mod::get()->setSavedValue<std::vector<bool>>("difficulty-array", { true, false, false, false, false, false });
-			else
-				verifyArray("difficulty-array", getFromSaveContainer("difficulty-array").as_array());
+			static_cast<void>(DataManager::setDefaultSafe<DMArrayKey::DIFFICULTY_ARRAY>());
+			static_cast<void>(DataManager::setDefaultSafe<DMArrayKey::DEMON_DIFFICULTY_ARRAY>());
+			static_cast<void>(DataManager::setDefaultSafe<DMArrayKey::SELECTED_LIST_ARRAY>());
 
+			gdListID = DataManager::setDefaultSafe<DMMiscKey::GD_LIST_ID>();
+			gameState = DataManager::setDefaultSafe<DMMiscKey::SAVE_DATA>();
 
-			if (!Mod::get()->hasSavedValue("demon-difficulty-array"))
-				Mod::get()->setSavedValue<std::vector<bool>>("demon-difficulty-array", { true, false, false, false, false });
-			else
-				verifyArray("demon-difficulty-array", saveContainer["demon-difficulty-array"].as_array());
+			if (gameState.levelID != 0)
+				currentPercentageGoal = gameState.levelPercentage;
 
-
-			if (!Mod::get()->hasSavedValue("selected-list-array"))
-				Mod::get()->setSavedValue<std::vector<bool>>("selected-list-array", { true, false, false, false });
-			else
-				verifyArray("selected-list-array", saveContainer["selected-list-array"].as_array());
-
-
-			if (!Mod::get()->hasSavedValue("gdListID"))
-				Mod::get()->setSavedValue<uint64_t>("gdListID", static_cast<uint64_t>(gdListID));
-			else
-				gdListID = static_cast<int>(Mod::get()->getSavedValue<uint64_t>("gdListID"));
-
-
-			previousDifficulty = getDifficultyFromSaveContainer("difficulty-array");
-			previousDemonDifficulty = getDifficultyFromSaveContainer("demon-difficulty-array");
+			previousDifficulty = getDifficultyFromSaveContainer<DMArrayKey::DIFFICULTY_ARRAY>();
+			previousDemonDifficulty = getDifficultyFromSaveContainer<DMArrayKey::DEMON_DIFFICULTY_ARRAY>();
 
 			m_hasInitManager = true;
 		}
 	}
 
-	// not using getSavedValue because i want a reference and not a copy
-	matjson::Value& getFromSaveContainer(const std::string_view key)
+	void addExclamationMark(CCMenuItemSpriteExtra* button = nullptr)
 	{
-		return Mod::get()->getSaveContainer().as_object()[key];
+		auto exclamationSprite = CCSprite::createWithSpriteFrameName(
+			"exMark_001.png"
+		);
+		exclamationSprite->setScale(.6f);
+		exclamationSprite->setID("exclamation-mark");
+		(button
+			? button
+			: rouletteLayer
+				->getParent()
+				->getChildByType<CreatorLayer>(0)
+				->getChildByID("center-left-menu"_spr)
+				->getChildByID("roulette-button"_spr)
+		)->addChild(exclamationSprite);
+
+		exclamationSprite->setLayoutOptions(
+			AnchorLayoutOptions::create()
+			->setAnchor(Anchor::TopRight)
+			->setOffset({ -5.f, -8.f })
+		);
 	}
 
-	bool getArrayState(matjson::Value& arr, std::size_t idx)
+	template <DMArrayKey key>
+	GJDifficulty getDifficultyFromSaveContainer()
 	{
-		return arr.as_array().at(idx).as<bool>();
-	}
+		int idx = rl::utils::getIndexOf(DataManager::get<key>().asVector(), true);
 
-	GJDifficulty getDifficultyFromSaveContainer(const std::string_view key)
-	{
-		int idx = rl::utils::getIndexOf(getFromSaveContainer(key).as_array(), true);
-
-		if (key == "demon-difficulty-array")
+		if constexpr (key == DMArrayKey::DEMON_DIFFICULTY_ARRAY)
 			return rl::constants::idx_to_demon_diff[idx];
 		else
 			return static_cast<GJDifficulty>(idx + 1);
@@ -135,27 +92,40 @@ public:
 
 	GJDifficulty getDifficultyFromSaveContainer()
 	{
-		int idx = rl::utils::getIndexOf(getFromSaveContainer("difficulty-array").as_array(), true);
+		int idx = rl::utils::getIndexOf(DataManager::get<DMArrayKey::DIFFICULTY_ARRAY>().asVector(), true);
 
 		if (auto difficulty = static_cast<GJDifficulty>(idx + 1); difficulty < GJDifficulty::Demon)
 			return difficulty;
 		else
 		{
-			int demonIdx = rl::utils::getIndexOf(getFromSaveContainer("demon-difficulty-array").as_array(), true);
+			int demonIdx = rl::utils::getIndexOf(DataManager::get<DMArrayKey::DEMON_DIFFICULTY_ARRAY>().asVector(), true);
 			return rl::constants::idx_to_demon_diff[demonIdx];
 		}
 	}
 
+	void saveState()
+	{
+		DataManager::set<DMMiscKey::SAVE_DATA>(gameState);
+	}
+
+	void resetState()
+	{
+		gameState.levelID = 0;
+		gameState.levelPercentage = 0;
+		gameState.levelPercentageGoal = 0;
+		gameState.numLevels = 0;
+		gameState.skipsUsed = 0;
+		gameState.hasReachedGoal = false;
+	}
+
 	void reset()
 	{
+		resetState();
+		saveState();
+
 		isPlaying = false;
 		isPaused = false;
-		hasFinishedPreviousLevel = false;
-		currentLevelID = 0;
-		currentLevelPercentage = 0;
-		levelPercentageGoal = 1;
-		skipsUsed = 0;
-		numLevels = 0;
+		currentPercentageGoal = 1;
 	}
 };